requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
detoxify==0.5.2
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
mss==10.0.0
networkx==3.4.2
numpy==2.2.5
openai==1.77.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1